removed ugly cleanup code in tests
Introduced factory for the app object and used that in the test
fixture. Introduced blueprints during that to keep the factory clean.
