+++ azure_billing/__init__.py
@@ -1,1 +1,3 @@
-from .main import app
+from .main import create_app
+
+app = create_app()
